Enum axmm::Backend

pub enum Backend {
    Linear {
        pa_va_offset: usize,
    },
    Alloc {
        populate: bool,
    },
}

A unified enum type for different memory mapping backends.

Currently, two backends are implemented:

Linear: used for linear mappings. The target physical frames are contiguous and their addresses should be known when creating the mapping.

Allocation: used in general, or for lazy mappings. The target physical frames are obtained from the global allocator.

Variants

Linear

Linear mapping backend.

The offset between the virtual address and the physical address is constant, which is specified by pa_va_offset. For example, the virtual address vaddr is mapped to the physical address vaddr - pa_va_offset.

Fields

pa_va_offset: usize

vaddr - paddr.

Alloc

Allocation mapping backend.

If populate is true, all physical frames are allocated when the mapping is created, and no page faults are triggered during the memory access. Otherwise, the physical frames are allocated on demand (by handling page faults).

Fields

populate: bool

Whether to populate the physical frames when creating the mapping.

Implementations

impl Backend

pub const fn new_alloc(populate: bool) -> Self

Creates a new allocation mapping backend.

impl Backend

pub const fn new_linear(pa_va_offset: usize) -> Self

Creates a new linear mapping backend.