axmm
Struct AddrSpace

pub struct AddrSpace { /* private fields */ }

The virtual memory address space.

Implementations

impl AddrSpace

pub const fn base(&self) -> VirtAddr

Returns the address space base.

pub const fn end(&self) -> VirtAddr

Returns the address space end.

pub fn size(&self) -> usize

Returns the address space size.

pub const fn page_table(&self) -> &PageTable

Returns the reference to the inner page table.

pub const fn page_table_root(&self) -> PhysAddr

Returns the root physical address of the inner page table.

pub fn contains_range(&self, start: VirtAddr, size: usize) -> bool

Checks if the address space contains the given address range.

pub fn map_linear( &mut self, start_vaddr: VirtAddr, start_paddr: PhysAddr, size: usize, flags: MappingFlags, ) -> AxResult

Add a new linear mapping.

The mapping is linear, i.e., start_vaddr is mapped to start_paddr, and start_vaddr + size is mapped to start_paddr + size.

The flags parameter indicates the mapping permissions and attributes.

Returns an error if the address range is out of the address space or not aligned.

pub fn unmap(&mut self, start: VirtAddr, size: usize) -> AxResult

Removes mappings within the specified virtual address range.

Returns an error if the address range is out of the address space or not aligned.

pub fn read(&self, start: VirtAddr, buf: &mut [u8]) -> AxResult

To read data from the address space.

Arguments

start - The start virtual address to read.

buf - The buffer to store the data.

pub fn write(&self, start: VirtAddr, buf: &[u8]) -> AxResult

To write data to the address space.

Arguments

start_vaddr - The start virtual address to write.

buf - The buffer to write to the address space.

pub fn protect( &mut self, start: VirtAddr, size: usize, flags: MappingFlags, ) -> AxResult

Updates mapping within the specified virtual address range.

Returns an error if the address range is out of the address space or not aligned.