Crate axmm

ArceOS memory management module.

Structs

AddrSpace

The virtual memory address space.

Enums

Backend

A unified enum type for different memory mapping backends.

Functions

init_memory_management

Initializes virtual memory management.

init_memory_management_secondary

Initializes kernel paging for secondary CPUs.

kernel_aspace

Returns the globally unique kernel address space.

kernel_page_table_root

Returns the root physical address of the kernel page table.

new_kernel_aspace

Creates a new address space for kernel itself.

new_user_aspace

Creates a new address space for user processes.